Time: 0 ms (100.00%) | Memory: 13.2 MB (97.98%) - LeetSync

+++ 1716-maximum-non-negative-product-in-a-matrix/maximum-non-negative-product-in-a-matrix.cpp
@@ -1,4 +1,4 @@
-int64_t dp[15][15][2];
+long long dp[15][15][2];
 class Solution {
 public:
     static int maxProductPath(vector<vector<int>>& grid) {
